+++ example/lib/main.dart
@@ -98,6 +98,7 @@ class _GameWheelPageState extends State<GameWheelPage> {
           //   width: 28,
           //   height: 47,
           // ),
+          indicatorAnimateStyle: 0,
           enableWeight: false,
           textStyle: const TextStyle(
             fontSize: 16,
@@ -109,6 +110,9 @@ class _GameWheelPageState extends State<GameWheelPage> {
           numberOfTurns: 10,
           onItemPressed: (index) {
             print('index: $index');
+          },
+          onItemSpinning: (index) {
+            // Play spin audio
           },
           onStartPressed: () {
             //Play start audio